add: support for decimal numbers

+++ src/main/java/davisantx/Expression.java
@@ -2,5 +2,4 @@ package davisantx;
 
 import java.util.ArrayList;
 
-public record Expression(ArrayList<Integer> numbers, ArrayList<Operations> operations) {
-}
+public record Expression(ArrayList<Double> numbers, ArrayList<Operations> operations) {}
